backend/requirements.txt
amqp==2.5.2
#ansible==2.9.2
anyjson==0.3.3
asgiref==3.2.3
autopep8==1.4.4
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
diff-match-patch==20181111
DingtalkChatbot==1.3.0
Django==2.0.4
django-auth-ldap==2.1.0
django-bulk-update==2.2.0
django-celery==3.3.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==2.4.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.2
django-import-export==2.0
django-rest-swagger==2.2.0
django-reversion==3.0.5
django-timezone-field==4.0
django-user-tracking==0.4.11
#django-xadmin==0.5.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-dynamic-fields==0.3.1
future==0.18.2
gunicorn==19.9.0
httplib2==0.15.0
idna==2.8
importlib-metadata==1.3.0
IPy==1.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.7
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.0.2
nmap==0.0.1
openapi-codec==1.3.2
paramiko==2.7.1
ply==3.11
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.2
PyNaCl==1.3.0
python-consul==1.1.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-gitlab==1.4.0
python-ldap==3.2.0
python-nmap==0.6.1
pytz==2019.3
PyYAML==5.3
redis==3.2.0
requests==2.22.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
tablib==1.0.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
xpinyin==0.5.6
zipp==0.6.0
pyOpenSSL==19.1.0
elasticsearch6==6.4.2